Refactor: Re-ordering the files and folders
Removing some of the useless files

Moving the RESful API into the corresponding directory
