--- public/editTrack.ejs
@@ -1,31 +1,0 @@
-<div class="container">
-    <h1>Edit Track</h1>
-    <hr>
-    <%if(error){%>
-        <p><%=error%></p>
-    <%}else{%>
-    <form method="post" enctype="multipart/form-data" id="trackEditForm" trackid="<%=track.id%>">
-        <p>Title: <input type="text" name="title" value="<%=track.title%>"></p>
-        <p>Tags: </p>
-        <div id="tags">
-            <div class="tags">
-                <%track.tags.map(tag => {%>
-                    <div onclick="removeTag(this)" trackid="<%=track.id%>" tagname="<%=tag%>"># <%=tag%></div>
-                <%})%>
-            </div>
-            <input type="text" id="tagInput" trackid="<%=track.id%>">
-        </div>
-        <br>
-        <p>Description: <textarea name="description" id="description"><%=track.description%></textarea></p>
-        <p>Cover Image: <input type="file" name="image"></p>
-        <%if(track.private){%>
-            <p>Private: <input type="checkbox" name="private" id="private" checked></p>
-        <%}else{%>
-            <p>Private: <input type="checkbox" name="private" id="private"></p>
-        <%}%>
-        <img src="<%=track.coverImage%>" alt="" class="coverImage">
-        <button class="btn btn-success" ng-controller="editTrack">Update</button>
-    </form>
-    <%}%>
-    <script src="/static/js/angular/tags.js"></script>
-</div>
