--- public/fave.ejs
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>HoovesSound | Music for bronies music for everyone</title>
-    <link rel="stylesheet" href="/static/css/fave.css">
-</head>
-<body>
-<div class="container">
-    <h1>Fave Tracks</h1>
-    <%if(tracks.tracks.length >= 1){%>
-    <div class="tracks">
-        <%tracks.tracks.forEach((track, index) => {%>
-        <%if(track !== null){%>
-        <div id="<%=track.id%>" style="background-image: url(<%=track.coverImage%>);" class="trackContainer">
-            <script>localPlayList.addTrack('<%-track.id%>')</script>
-            <div class="playPuaseButton material-icons" fullname="<%=track.author.fullName%>" username="<%=track.author.username%>" title="<%=track.title%>" trackid="<%=track.id%>" onclick="playMusic(this)">play_arrow</div>
-            <a href="/track/<%=track.author.username%>/<%=track.title%>" ng-controller="homeTrackLink" ng-click="homeTrackLink($event); $event.stopPropagation();" fullname="<%=track.author.fullName%>" username="<%=track.author.username%>" title="<%=track.title%>" class="trackName"><%=track.author.fullName%> - <%=track.title%></a>
-        </div>
-        <%}%>
-        <%});%>
-
-    </div>
-    <%}else{%>
-    <p>You didn't fave any tracks yet</p>
-    <%}%>
-</div>
-</body>
-</html>
