--- public/includes/adsense.ejs
@@ -1,7 +1,0 @@
-<script async src="//pagead2.googlesyndication.com/pagead/js/adsbygoogle.js"></script>
-<script>
-    (adsbygoogle = window.adsbygoogle || []).push({
-        google_ad_client: "ca-pub-5677639700264267",
-        enable_page_level_ads: true
-    });
-</script>
