--- public/includes/artwork.ejs
@@ -1,13 +1,0 @@
-<div class="bgAuthor">
-    Artwork by: <br>
-    <a href="<%=background.author.link%>" target="_blank"><%=background.author.name%></a>
-</div>
-<style>
-    html{
-        background: url(<%=background.image%>);
-        background-repeat: no-repeat;
-        background-size: cover;
-        width: 100vw;
-        height: 100vh;
-    }
-</style>
