--- public/includes/nav.ejs
@@ -1,59 +1,0 @@
-<link rel="stylesheet" href="/static/css/nav.css">
-
-<nav class="navbar navbar-inverse navbar-static-top">
-    <div class="container-fluid">
-        <div class="navbar-header">
-            <button type="button" class="navbar-toggle" data-toggle="collapse" data-target="#hsNavBar">
-                <span class="icon-bar"></span>
-                <span class="icon-bar"></span>
-                <span class="icon-bar"></span>
-            </button>
-            <a class="navbar-brand" href="#">HoovesSound</a>
-        </div>
-
-        <div class="collapse navbar-collapse" id="hsNavBar">
-            <ul class="nav navbar-nav">
-
-
-                <li><a href="/home">Home <span class="sr-only"></span></a></li>
-                <li><a href="/me/fave">Fave <span class="sr-only"></span></a></li>
-                <form class="navbar-form navbar-left" role="search" id="searchForm" ng-controller="search">
-                <div class="form-group">
-                    <input type="text" class="form-control" placeholder="Search" name="query">
-                    <button type="submit" class="btn btn-default">search</button>
-                </div>
-                </form>
-
-            </ul>
-
-            <!--Right side-->
-
-            <ul class="nav navbar-nav navbar-right">
-                <li class="dropdown">
-                    <a class="dropdown-toggle" data-toggle="dropdown" href="#">
-                        <img src="<%=loginUser.icon%>" alt="" id="userProfilePicture">
-                        <span class="caret"></span>
-                    </a>
-                    <ul class="dropdown-menu">
-                        <li><a href="/me">Profile</a></li>
-                        <li><a href="/settings">Settings</a></li>
-                        <li><a href="/me/apps">My Apps</a></li>
-                    </ul>
-                </li>
-            </ul>
-
-            <ul class="nav navbar-nav navbar-right">
-                <li><a href="/notification/" class="notificationButton material-icons">notifications</a></li>
-            </ul>
-
-            <ul class="nav navbar-nav navbar-right">
-                <li><a href="/upload">Upload</a></li>
-            </ul>
-
-
-
-        </div>
-    </div>
-</nav>
-
-<script src="/static/js/search.js"></script>
