--- public/profile.ejs
@@ -1,49 +1,0 @@
-<div class="container">
-    <link rel="stylesheet" href="/static/css/profile.css">
-    <%if(profile === null){%>
-    <div class="container">
-        <p>There not such a user :/</p>
-    </div>
-    <%}else{%>
-
-    <div class="container">
-        <div id="userInfo" style="background-image: url('<%=profile.banner%>')">
-            <img src="<%=profile.icon%>" alt="" class="profilePicture">
-            <div id="roles">
-                <%profile.roles.forEach(role => {%>
-                <%const title = Object.keys(role)[0];%>
-                <%const fsIcon = role[title];%>
-                <div class="role <%=title%> fa fa-<%=fsIcon%>" title="<%=title%>"></div>
-                <%});%>
-            </div>
-            <h1 id="username"><%=profile.fullName%></h1>
-            <%if(profile.username === loginUser.username){%>
-            <a class="btn btn-success" href="/settings">Settings</a>
-            <%}%>
-        </div>
-
-        <hr>
-        <h1 class="title">Users' tracks</h1>
-        <%if(tracks.length <= 0){%>
-        <p>Looks like this user never upload an track yet :/</p>
-        <%}else{%>
-        <%tracks.forEach(track => {%>
-        <div class="tracks">
-            <div id="<%=track.id%>" style="background-image: url(<%=track.coverImage%>);" class="trackContainer">
-                <script>localPlayList.addTrack('<%-track.id%>')</script>
-                <div class="playPuaseButton material-icons" fullname="<%=track.author.fullName%>" username="<%=track.author.username%>" title="<%=track.title%>" trackId="<%=track.id%>" onclick="playMusic(this)">play_arrow</div>
-                <a href="/track/<%=track.id%>" ng-controller="homeTrackLink" ng-click="homeTrackLink($event); $event.stopPropagation();" fullname="<%=track.author.fullName%>" username="<%=track.author.username%>" title="<%=track.title%>" class="trackName" trackId="<%=track.id%>"><%=track.author.fullName%> - <%=track.title%></a>
-            </div>
-        </div>
-        <%})%>
-        <%}%>
-    </div>
-    <script>
-        <%if(tracks.length > 0){%>
-        if(audio.paused){
-            audio.src = "<%=full_address%>/api/listen/<%=tracks[0]._id%>?token=<%=token%>";
-        }
-        <%}%>
-    </script>
-    <%}%>
-</div>
