--- public/search.ejs
@@ -1,35 +1,0 @@
-<link rel="stylesheet" href="/static/css/search.css">
-<div class="container">
-    <h1>Search</h1>
-    <%if(users){%>
-    <%users.forEach(user => {%>
-    <div class="user">
-        <a href="/@<%=user.username%>">
-            <img src="<%=user.icon%>" alt="" class="icon">
-            <%=user.fullName%>
-        </a>
-    </div>
-    <%})%>
-    <%}%>
-
-    <%if(tracks.length >= 1){%>
-    <%tracks.forEach((track) => {%>
-    <div style="background-image: url(<%=track.coverImage%>);" class="trackContainer">
-        <script>localPlayList.addTrack('<%-track.id%>')</script>
-        <div class="playPuaseButton material-icons" fullname="<%=track.author.fullName%>" username="<%=track.author.username%>" title="<%=track.title%>" trackId="<%=track.id%>" onclick="playMusic(this)">play_arrow</div>
-                <a href="/track/<%=track.id%>" ng-controller="homeTrackLink" ng-click="homeTrackLink($event); $event.stopPropagation();" fullname="<%=track.author.fullName%>" username="<%=track.author.username%>" title="<%=track.title%>" class="trackName" trackId="<%=track.id%>"><%=track.author.fullName%> - <%=track.title%></a>
-    </div>
-    <%});%>
-    <%}%>
-
-    <%if(tags){%>
-        <spam># <%=tags.name%></spam>
-    <%}%>
-
-    <%if(tracks.length <= 0 && users.length <= 0 && !tags){%>
-    <div class="emptyResult">
-        <h3>~ Ahhh ~</h3>
-        <h4>We can't find you anything here</h4>
-    </div>
-    <%}%>
-</div>
