--- public/settings.ejs
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>HoovesSound | Settings</title>
-    <link rel="stylesheet" href="/static/css/settings.css">
-</head>
-<body>
-    <div class="container">
-        <%if(error){%>
-            <div class="error">
-                <%=error%>
-            </div>
-        <%}else{%>
-            <div class="error"></div>
-        <%}%>
-        <h1>Settings</h1>
-
-        <div id="userInfo" style="background-image: url('<%=loginUser.banner%>')">
-
-            <div id="profilePicture">
-                <img src="<%=loginUser.icon%>" alt="" class="icon">
-                <input type="file" name="icon" id="iconInput">
-            </div>
-
-            <h1 id="fullname" ><input type="text" value="<%=loginUser.fullName%>"></h1>
-        </div>
-
-    </div>
-    <script src="/static/js/settings.js"></script>
-</body>
-</html>
