--- public/track.ejs
@@ -1,83 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
-    <link rel="stylesheet" href="/static/css/app.css">
-    <link rel="stylesheet" href="/static/css/track.css">
-    <%if(track !== null){%>
-        <title>HoovesSound | <%=track.title%></title>
-    <%}else{%>
-        <title>HoovesSound | No Tracks :/</title>
-    <%}%>
-</head>
-<body>
-    <div class="container">
-        <%if(typeof error !== 'undefined'){%>
-        <h1>Error</h1>
-        <p><%=error%></p>
-        <%}else{%>
-        <%if(track !== null){%>
-        <%if(track.author.username === loginUser.username){%>
-        <a href="/track/<%=track.id%>/edit" class="btn btn-default" ng-controller="toEditLink" ng-click="toEditLink($event); $event.stopPropagation();" username="<%=track.author.username%>" title="<%=track.title%>" trackId="<%=track.id%>">Edit</a>
-        <%}%>
-        <div class="fa fa-heart fave <%=isFave%>"></div>
-        <h1 id="title"><%=track.title%></h1>
-        <div id="tags">
-        
-        </div>
-        <p>By <a href="@<%=track.author.username%>">@<%=track.author.username%></a></p>
-        <div class="playPuaseButton material-icons" fullname="<%=track.author.fullName%>" username="<%=track.author.username%>" title="<%=track.title%>" trackid="<%=track.id%>" onclick="playMusic(this)">play_arrow</div><br>
-        <a class="share btn btn-info">Widget</a>
-        <div class="tags">
-            <%track.tags.map(tag => {%>
-            <div ng-click="gotoTag" trackid="<%=track.id%>" tagname="<%=tag%>"># <%=tag%></div>
-            <%})%>
-        </div>
-        <input type="text" class="shareURL hide" value='<iframe src="<%=full_address%>/api/widget/<%=track.id%>"></iframe>' readonly>
-
-        <img src="<%=track.coverImage%>" alt="" class="coverImage">
-
-        <hr>
-
-        <h1>Description</h1>
-        <div id="description">
-            <p><%-track.description%></p>
-        </div>
-
-        <hr>
-        <h1>Comments</h1>
-        <div id="commentSession">
-            <textarea name="comment" id="input"></textarea>
-            <button type="submit" id="post">Post</button>
-            <div id="comments">
-                <%if(comments){%>
-                <%comments.forEach(comment => {%>
-                <div class="comment">
-                    <a href="/user/<%=comment.author.username%>">
-                        <%=comment.author.fullName%>
-                    </a>
-                    <p><%-comment.comment%></p>
-                </div>
-                <%})%>
-                <%}else{%>
-                <h1>No body have comment yet, will ya?</h1>
-                <%}%>
-            </div>
-        </div>
-        <%}else{%>
-        <h1 id="notracks">Didn't have any tracks yet</h1>
-        <%}%>
-        <script>
-            var trackid = "<%=track.id%>";
-        </script>
-        <script src="/static/js/comment.js"></script>
-        <script src="/static/js/track.js"></script>
-        <%}%>
-    </div>
-
-</body>
-</html>
