--- public/tracks.ejs
@@ -1,20 +1,0 @@
-<div class="container">
-    <h1 class="title">Latest Tracks</h1>
-    <%if(tracks.length >= 1){%>
-    <div class="tracks">
-        <%tracks.forEach((track) => {%>
-        <div style="background-image: url(<%=track.coverImage%>);" class="trackContainer">
-            <script>localPlayList.addTrack('<%-track.id%>')</script>
-            <div class="playPuaseButton material-icons" fullname="<%=track.author.fullName%>" username="<%=track.author.username%>" title="<%=track.title%>" trackId="<%=track.id%>" onclick="playMusic(this)">play_arrow</div>
-            <a href="/track/<%=track.id%>" ng-controller="homeTrackLink" ng-click="homeTrackLink($event); $event.stopPropagation();" fullname="<%=track.author.fullName%>" username="<%=track.author.username%>" title="<%=track.title%>" class="trackName" trackId="<%=track.id%>"><%=track.author.fullName%> - <%=track.title%></a>
-        </div>
-        <%});%>
-    </div>
-    <%}else{%>
-    <p>No one have ever upload an track yet, will ya?</p>
-    <%}%>
-    <div class="btn btn-default loadMore" ng-controller="getMoreContent" ng-click="getMoreContent($event); $event.stopPropagation();">Load More</div>
-    <script>
-        var offset = <%=offset%>;
-    </script>
-</div>
