--- src/router/view/upload.js
@@ -1,32 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const rp = require('request-promise');
-const fullurl = require('fullurl');
-const Users = require('../../../schema/Users');
-
-router.get('/', (req, res) => {
-    if(!req.cookies['oauth-token']){
-        res.end('Access denied');
-    }else{
-        const token = req.cookies['oauth-token'];
-        const full_address = req.protocol + "://" + req.headers.host;
-
-        Users.findOne({
-            token,
-        }).then(user => {
-            if(user === null){
-                res.end('Access denied');
-            }else{
-                res.render('upload', {
-                    loginUser: user,
-                    error: null,
-                    token,
-                });
-            }
-        }).catch(error => {
-            console.log(error)
-        });
-    }
-});
-
-module.exports = router;
